server/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-nested-routers==0.93.5
exceptiongroup==1.2.0
Faker==22.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-watch==4.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
